Garmin: Remove File Name from File Data.
Remove the file name from the beginning of the file data buffer. This is
in preparation to a move to shift the .FIT file importer to 'Import log
files'.
The file name in the buffer does not seem to be used for anything, and
it seems to be wrong to have this added only for the parser to then have
to remove this non-file-content again first thing.
Fingerprinting has been confirmed as working after this change.

Signed-off-by: Michael Keller <github@ike.ch>